requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
spotipy
Werkzeug
PyJWT
Flask-Migrate
